move package for distribution plan
